+++ framework/db/ActiveQuery.php
@@ -67,7 +67,7 @@ class ActiveQuery extends Query implements ActiveQueryInterface
 		$command = $this->createCommand($db);
 		$rows = $command->queryAll();
 		if (!empty($rows)) {
-			$models = $this->createModels($rows, false);
+			$models = $this->createModels($rows);
 			if (!empty($this->join) && $this->indexBy === null) {
 				$models = $this->removeDuplicatedModels($models);
 			}
@@ -144,7 +144,7 @@ class ActiveQuery extends Query implements ActiveQueryInterface
 			} else {
 				/** @var ActiveRecord $class */
 				$class = $this->modelClass;
-				$model = $class::create($row, false);
+				$model = $class::create($row);
 			}
 			if (!empty($this->with)) {
 				$models = [$model];