+++ extensions/elasticsearch/CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 elasticsearch extension Change Log
 2.0.0 beta under development
 ----------------------------
 
-- Bug #1993: afterFind event in AR is now called after relations have been populated (cebe)
+- Bug #1993: afterFind event in AR is now called after relations have been populated (cebe, creocoder)
 - Enh #1382: Added a debug toolbar panel for elasticsearch (cebe)
 - Enh #1765: Added support for primary key path mapping, pk can now be part of the attributes when mapping is defined (cebe)
 - Chg #1765: Changed handling of ActiveRecord primary keys, removed getId(), use getPrimaryKey() instead (cebe)